Add files via upload
A solar cell built of ultra-thin amorphous silicon and high-quality mono-crystalline silicon is known as a hetero-junction intrinsic thin film. It has a pyramid surface on the front that increases sunlight absorption. The operating environment has a significant impact on the performance of hetero-junction intrinsic thin-film photovoltaic modules with real I–V (current-voltage) characteristics. Changes in the environment have a significant impact on solar irradiation. Clouds also have a significant impact on the solar irradiation that a PV cell receives. In this project, we will use the Random Forest Regression machine learning algorithm to investigate the effects of sudden changes in environmental conditions on power output and module temperature of an HIT (Heterojunction with Intrinsic Thin Layer) module, where irradiance, temperature, and module efficiency parameters are taken into account when designing modules. The algorithm’s output will be studied to gain a better understanding of performance variations as well as the behavior of the power output and module temperature when subjected to random influences induced by various environmental variables. The suggested algorithm is not restricted to a certain module technology or geographic location.